Resto Druid
My own version with some tweaks.
Firstly it will autofocus if you dont have a focus, Basically it will
make your targetstarget to focus.
It will cast mushroom on focus target if it is 10 yards from focus and
is not moving.
IT will also use Filler reju count as max number of reju thats are up.
So use the count so you wont spamm reju. Set it high if you weant to
spam rejuvs.

Next it will only use Barkskin if you are badly debuffed. So set
barkskin on low HP and it will be fine