Nightly
System - Adjusted cast logic again to better differentiate when to use a
castGround at target spell
Vengeance - Fixed Sigil of Silence and Sigil of Misery and Infernal
Strike
Guardian - Fixed issue with mauling friendly targets, remember friends
are not food.
Resto Druid has not been forgotten about and is still in the works, in
addition to more Blood DK work.