Make queue casting great again
For some reason the queuecast function was placed after rotation in core. I think that if I queue up a cast, I want it to cast it next time.
Also it wasn't using the castSpell function, so would sometimes go in a loop, spamming the same action over and over.